Don't skip over recordings that are still tuning.
Fix a bug where the scheduler sometimes skips over recordings in the
rsTuning state.  This doesn't matter normally as the recording usually
transitiions to the rsRecording state quickly.  When there is a tuning
problem, however, this bug defeats the tuning timeout check in the
scheduler and the recording gets stuck in the tuning state.  This fix
makes sure the tuning check is always done.

(cherry picked from commit f455e491239cc128200db85c90dcab7ea15044c3)